First draft of new parallelism style.
Reworked the calculateClusteredMIPScores to take advantage of new
poolInit and ETMIPWorker functions that utilize queues.  Changes to
poolInit function include the addition of other shared variables (aaDict
and others) as well as the queues. Additions to ETMIPWorker set up to
utilize queues. One to track the k's that need to be worked on, another
to track the subalignments, and one to track results. The method
preferentially pulls from queues further down the pipeline
(sub-alignments) and only pulls from the first queue if there is nothing
else to work on. Biggest potential hiccups are integrating results after
threading and the order of treeOrdering operations depending on how
cluster constants are pulled from the queue.