Troubleshooting initial attempt to improve parallelism.
Running with one processor, I am trying to troubleshoot simple errors in
the code. Caught an incorrect usage of setTreeOrdering in
PerformAnalysis. It still passed the cacheDir variable which was removed
because the variable was no longer used by the method. Corrected error
in ETMIPC method etMIPWorkerTemp where I used += on an unassigned
dictionary element.