Added multiprocessing to the produceFinalFigures method.
This required the creation of two new non-class methods in the ETMIPC
code, one to govern the initialization of pool processes and the other
for the processes to execute.  These methods enable the creation and
filling of one queue of jobs to be done and a second queue with actual
functions to call to create output files.  To make previous methods
available to these functions it was necessary to move several functions
from the class definition, including: plotAUC, heatmapPlot, surfacePlot,
writeOutClusteringResults, and writeOutClusterScoring. Since these were
class methods it was necessary to remove the self references and add
additional inputs to each method.