Corrected and refined initial set of calls to the ET-MIp binary. The code has been tested as is and works on one of the lab workstations using the specified environment.
The environment.yml has been updated to reflect the addition of the dotenv package. The .gitigore file has been updated so that the .env file will not be tracked in the repository, while the .env_template has been added so that users know which variables need to be specified. The ETMIPWrapper has been corrected and tested so that the code runs as specified. A method to remove output files has been added to clean up during execution.