First attempt to include clustering z-score code.
This code was provided by Eunna Huh and Ilya Novikov and comes from Python code by Rhonald Lua and C code by Angela Wilkins. The appropraite citations are in the documentation. I have done my best to adapt these to the current framework of the ContactScorer class.
TODO I believe there are several improvements which could be made to the speed of this method vis a vis storing values that do not need to be recomputed every time.
TODO I also think it might be important to re-evaluate the naming of this class it may be more correct to call it co-evolution or covariance scorer.