Updated documentation for _clustering_z_score and implemented score_clustering_of_contact_predictions.
Provided first implementation of score_clustering_of_contact_predictions. This first implementation attempts to add pairs of residues in sorted order and calculate their z-score. If a pair does not map to the pdb, the new residues added by the pair are removed from the set of residues of interest before the next pair is analyzed so that later calculations will not be affected by the bad residue(s). There are likely better ways to implement this so it may need to be updated in the future.