Adding DCA code and wrapper to the project.
This is being added to automate the comparison of cET-MIp, ET-MIp, and DCA. The .env_template has been updated to track the PROJECT_PATH, which will be used to find cmd_line_GausDCA.jl. This file is a set of calls to the DCA code written in julia, which takes in the path to an alignment and a path at which to write out the covariance scores. A DCAWrapper has been written in python to call this julia code and to import the scores.
TODO: Test new environment including the julia installation and updated python packages.