Updated environment to include julia.
During this update I also changed the name of the environment from
pyETMIPC to cET-MIp to reflect the naming of the tool correctly. I also
updated several of the python packages the changes include the following:
python-dotenv: 0.8.2 -> 0.9.1
biopython: 1.69 -> 1.72
seaborn: 0.7.1 -> 0.9.0
scipy: 0.19.1 -> 1.1.0
pandas: 0.20.3 -> 0.23.4
numpy: 1.13.1 -> 1.13.3
The only change of these which was required as the numpy update. This
was required because the c depndencies for julia were in conflict with
the c dependencies for numpy. I tested the output of the code on the
master branch at this point with the original environment and the
current code in this (reducing_memory_footprint) branch with the old and
new environment and found only one change. In the files which record the
auc curve data, the residue distances no longer have very long decimal
point precision, but instead are cutoff at the 4th decimal point. This
change in precision/rounding does not affect any of the scoring for the
proteins I used to test the code (7hvpA, 1bolA, 2ysdA).