On going work to correct errors in z-scoring methods.
Set up the 23TestProteinInitialEvaluation script to run on the new code
structure and outlined it with DCA as the model. Ran into several
hiccups. At this point AUC and Precision scores seem to be working. The
clustering z-scores are not being computed correctly so at this point the
code does not run all the way through.