Reverse engineering and troubleshooting z-scores.
Took Eunna/Rhonald's code and got it to run. Was able to get mine to run with the same behavior for some/all examples (complete test not yet finished).
TODO: Find out why in Rhonald's code the square of the distance not the distance itself is checked against the cutoff.
TODO: Complete comparison to z-scores computed with old method.