After much troubleshooting the z-scoring code has been corrected.
I used Eunna and Rhonald's code to compare to my results. The take away for me is the following:
Using i and j as variable names leads to problems if you accidentally type i,i and meant i,j but read over it many times.
If there is a default cutoff and you change it your results will not match precomputed results (I had started using 8 Angstroms as my clustering cutoff and this did not match the results from Eunna at all who was using the default 4 Angstroms)
If performing matrix math to increase speed think carefully about whether the diagonal should be set to 0 or not, and if you are using an upper or lower triangle whether this will include the diagonal or not.