Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
This file is very incomplete and requires a lot more work. Current effort was to match the init function to the init functions in the DCA and ETMIP wrapper classes. Also wrote a calculate_scores method to match the other to classes. THere are lots of loose ends. Some suggestions would be to move AUC computation out of this class (proabably into scorer). I would also suggest moving figure creation out of this class (also probably into the scorer). Need to finish cleaning up this code.