Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Trying to preserve PerformAnalysis behavior given new class frameworks. Changed it so that PerformAnalysis does not cd into the create directory, now filenames are passed explictly to methods that will write out files. Compared the import_alignment method in ETMIPC to the first part of analyze_alignment in PerformAnalysis, and they seem to be equal.