Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Tested calculate_scores vs old perform analysis calls. Added a method to contact scorer to evaluate an entire set of results from a predictor and another to evaluation individual sets of predictions (a matrix). These methods also produce a final evaluation file which is similar to that produced as part of the cET-MIp output before.