Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Cleaned up commented code in ETMIPC as well as redundant checking of variable contents instead of just passing the status of low_mem to different processes. Changed loading of matrices to just rely on the stored file paths instead of needing all of the variable and name descriptions to find the desired data in the output directory.