Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Updated getter functions to work correctly. Used them within ETMIPC where appropriate. Unfortunately many of the places where this would have been convenient are in the multiprocessed methods and are there for not handled correctly. Perhaps this can be solved by writing non-class methods which can also be called within in the class.