Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Updated the ContactScorer class to reduce computation for 23TestProteinInitialEvaluation when re-running analysis from a partial previous run. Changed surface_plot, heatmap_plot, plot_zscores, and plot_auc so that plots are not generated again if they already exist. Updated score_clustering_of_contact_predictions, evaluate_predictions, and evaluate_predictor so that data which can be re-used in different code blocks is not recomputed and so that data are passed between methods which may need them for filewriting or plotting have easy access to them.