Working to make ETMIPC class compatible with DCA and ETMIP wrapper classes.
Working on a new version of the set_tree_ordering function which is not yet complete! Also added a new method which generates subalignments based on two positions (as opposed to using SeqRecord IDs. Updated the heatmap_plot function to make saving optional, to allow the passing of the amino acid dictionary, and to return the generated Axes object when the heatmap is plotted.