Adding to SeqAlignment test.
Made two changes to SeqAlignment. First, I made re_label_clusters a static class method and also encapsulated it. This method still needs to be tested. Second, I added several parameters to set_tree_ordering: tree_depth, clustering, and clustering_args. Tree_depth is a stopping criteria for set_tree_ordering it is meant to limit the runtime of this method by allowing only the required branches to be specified. Clustering is backed by a dictionary of functions within the method call which allows the clustering used to build the tree to specified. Clustering_args is a dictionary meant for passing additional arguments to any clustering method requested for tree building.