Adding to SeqAlignment test.
Updated visualize_tree in the SeqAlignment class so that the paths were not hard coded and to ensure that the tree_order was followed instead of the seq_order. Wrote but have not performed test for visualize_tree. Eventually it might be good to visualize an actual tree (which probably comes with checking for conforming to expected rules) instead of the heat mapping method I am using at present.