Reworking ETMIPC and testing.
Tested _generate_sub_alignments, _generate_sub_alignment, and _pool_init_sub_aln. Turned out even if the fucntions were static they could not be passed to the pool because they are not picklable if they are in the class definition. So I moved generate_sub_alignment and pool_init_sub_aln out of the class definition and removed their staticmethod decorator. The tests now pass, and I added one to explicitely check the multiprocessing case.