Updating Documentation and Testing: SeqAlignment
Corrected compute_distance_matrix documentation. Also updated the method so it can take file names with or without the .npz suffix (changes were only needed for my loading code as np handles this already). This computation using the biopython distance calculator is very slow, I wonder if it would be worth submitting my solution for this calculation to the public project, I am not sure because I have never done something like that.