Updating Documentation and Testing: ETMIPC
Completed testing of ETMIPC. When all tests are run in order an error happens for test_exists_single_matrix. This may suggest that files are not properly cleared in one of the other tests, since the test completes successfully when run on its own. This suit of tests takes ~2 hours to run, there must be better ways to perform these tests, perhaps with setUpClass and tearDownClass.