Adding UPGMA tree building
First draft of a method to use UPGMA trees instead of random or hierarchical trees. Have not implemented tree traversal yet to retrieve sequences for a given cluster. Need to check how UPGMA was used in previous lab tools at the moment it uses the identity based distance metric which may not be correct.