Functional UPGMA tree traversal.
The new code successfully traverses the UPGMA tree to the desired depth. There are significant inefficiencies which I would like to improve. First, if a node in the traversal at a given level is not changing (was not the one that split) all its data can be copied from the level above.