Updating SeqAlignment and addition of UPGMA tree to set tree ordering.
Changes have been made and still need to be finalized for how distance matrices are used by this class. Instead of storing a np.array a Bio.Phylo.TreeConstruction.DistanceMatrix is stored because it can easily be converted to a np.array, while it is comparatively expensive to convert an np.array back to a DistanceMatrix object. This change still needs to be properly evaluated over all methods in the class, currently things seem to work correctly, but it would likely be more correct to store the preferred distance model as a class attribute instead of leaving it as an optional variable in so many methods.

UPGMA tree construction has been added, along with a simple test. A more complicated test will be implemented as part of the set_tree_ordering method test. The current implementation includes the creation of a dictionary of assignments which tracks which sequences are assigned to which clusters at each level in the tree. This allows for behavior similar to that of the agglomerative clustering method where the full tree is computed once and cluster labels are returned for each level where a request is made.