Updating PerformAnalysis to new code base
With this commit, PerformAnalysis.py again behaves correctly as the commandline interface for a single analysis with ETMIPC and an evaluation using ContactScorer. To make some old behaviors available again ContactScorer was updated so that either a path or a SeqAlignment object could be provided for the query_alignment variable, and either a path or a PDBReference object could be provided for the pdb_reference variable at initialization.