Updating SeqAlignment and testing
Split code for traversing a rooted Bio.Phylo.BaseTree from the _upgma_tree method so it can be used in multiple methods. Added a new method which should be able to read in a custom tree and return specified clusters from it. There are currently issues with building random trees because their structure does not follow the expectation for a rooted bifurcating tree, which breaks the _re_label_clusters method behavior. This still needs to be fixed and that correction needs to be tested as does reading in custom trees.