Correcting SeqAlignment tree traversal algorithm
The tree traversal algorithm was causing strange behavior for the _custom_tree method. This was due to the fact that I had built a node look up table for when nodes get reused which relied on node names. Node names are not necessarily unique (the upgma algorithm seems to generate unique names but there is no guarantee for newick formatted trees read in from disk). It turns out that Bio.Phylo Clade objects are hashable so they can be used as the dictionary keys directly, which keeps us from needing to use the names as a proxy. Switching to Clade hashing instead of name hashing did not change the behavior of the method as tested in the SeqAlignment set_tree_ordering test. It did correct the erroneous cluster assignments found in practice when using custom trees.