Runnable version of Parameterization script
This script has been updated to include print statements to track where in the execution we are. It also now has a full set of initial parameters to test which were incomplete after I had to revert the code yesterday after attempting a refactor which did not work correctly. This could still use improvement as it is quite slow. I think there is additional serialization which can be done in the parameterization script and multiprocessing which can be achieved in the ContactScorer methods.