Correcting Race Condition in ETMIPC
There was a race condition created by accessing the disk and checking for the existence of a directory in _score_clusters. When spawning many processes in the pool to run mip_score (I observed the behavior regularly with 11 processes) some processes will fail when they execute save_single_matrix. As far as I can tell the reason for this is that two processes processing clusters for the same branch level (i.e. clusters 4 and 5 at branching level 8) check for the presence of the directory where they will write the file. Both processes find that the directory does not exist. Both then try to write the directory but one wins out, so the second one encounters an OSError when trying to create the directory which it saw was not present yet, but was already created by the competing process. I have corrected this in several ways. I have preemptively created the expected directories before the pool is initialized (in _score_clusters). I have also surrounded any call to makedirs or mkdir that exists in a function which might run in a multiprocess setting with a try catch (even if it is preceded by an if statement checking for the existence of the directory).