Adding alignment write out and visualization back into ETMIPC
When clustering scores are being computed the ETMIPC code will now write out the sub-alignments as .fa alignment files and heatmap a visualizatoin of the alignment like it did before. This is not multiprocessed or sped up in any other way, but it takes very little time so hopefully it is never a big slow down.