Multiprocessing scw z-score calculation
I think further speedups must be available but right now I have multiprocessed the calculation of w2_ave_sub so that it can be computed more quickly when multiple processors are made available. The multiprocessing farms out the calculation of the w2_ave_sub scores for each position in the protein to a processing pool and returns the values for each of the three cases for that position. When the pool completes all of these values are combined together to give the final answer.