Improving speed of calculate_score_and_coverage
This method has been improved my replacing the double for loop used to compute the rank of the different scores with the scipy.stats.rankdata method. This significantly decreases the time to compute coverage. I had initially tried to vectorize the for loops which worked but was quite expensive time wise and memory wise (my work station ran out of memory for one of our test proteins). The idea to use rankdata came from Chih-Hsu Lin. After some testing I found that method='max' resulted in the same rank as what we had previously computed.