Attempting to speed up SCW Z-Scoring over the full structure
I have changed how the new multiprocessed code works, relying less on DataFrames and additional sorting steps, and more on dictionaries and lists. I have also made the change that unmappable pairs will be written at the end of the files as opposed to with their sorted scores earlier on. I think this should be acceptable since they do not influence z-scoring and are still viewable.