Updating ContactScorer to reduce repetitive work.
Scoring a predictor significantly duplicated the work by scoring the final scores and the coverage. Coverage and final score should theoretically be identically ranked (which is the basis of all the tests). I tested the scores for the evaluations between the final score and coverage for all k's in one protein's evaluation and they were exactly equivalent for all tests. So all coverage scoring is being removed. This will also significantly reduce the time for evaluations.