Added configuration logging to PerformAnalysis
A file is now written out with the settings for each of the variables available on the commandline. I hope this will make it easier to track experiments and to troubleshoot when something goes wrong in a larger experimental script (i.e. an error in parameterization could be checked by running PerformAnalysis with the settings described for the failed experiment).