SCW Z-Scoring corrected
Up to now I have made the distance cutoff for the Structural Cluster Weighting Z-Score variable so that it can be set by others. This has led to results which are not actually comparable with previous work in the lab (because I used the cut off for structural contacts used by CASP and most of the covariation/contact prediction field, i.e. 8.0 Angstroms). I have hard coded the distance cutoff to 4.0 Angstroms as seen in all previous versions of this code and the publications on it.