Added locking to the sub-alignment generation.
This was done because writing out heatmap of the alignment was made a defualt behavior and there is an error when trying to open some matplotlib resources across multiple processes. Locking just before generating the image avoids these errors. There are also two new version of the MIp computation, the first one (New MIp in comments) attempts to compensate the errors in the MIp calculation for single sequence alignment or high identity alignments by checking conditions in the computation and using default values. The second version (Newest MIp in the comments) is not actually an MIp calculation but an adjusted mutual information score which has its own errors which I had not fully troubleshot but which does lead to more ET like scoring as far as I am concerned.