Several updates to SeqAlignment
Updated compute_distance_model so that it uses my faster implementation of the identity distance computation again, this is special cased and used only when 'identity' is specified as the distance model. Wrote new private method, _convert_array_to_distance_matrix which allows a numpy array to be converted to the DistanceMatrix expected by Bio.Phylo.DistanceTree constructors. I use numpy arrays for fast accessing and vectorization which is not as easy to achieve with python lists which is essentially what the DistanceMatrix uses. generate_sub_alignment now also subsets the distance matrix when creating a new alignment, this takes advantage of the new _conert_array_to_distance_matrix function. _alignment_to_num now computes its own mapping dictionary based on the protein alphabet and the expected gap symbol so it is no longer needed as an argument to the method. heatmap_plot now creates the desired filename and checks if the file already exists, if so it is not created again. I also updated the call to _alignment_to_num() and the naming of columns when creating the dataframe which is turned into a heatmap.