New version of parameterization sweep I had been performing.
This is tailored to only check the factors of the highest interesting and which allow for comparison to ET. That means only identity (and the custom identity used by ET) are used for distance metrics. The only trees assessed are the ones built by agglomerative clustering with euclidean distance and ward linkage (one of the innovations of this new version of ET-MIp), upgma as implemented by Biopython (which is what the literature on ET says was being used), and the custom upgma implementation from the ET code base.