Updated SeqAlignment after reading the etc_read_msf.c file.
The class now includes a marked attribute to mirror the marked status for sequences tracked by the etc codebase. This is reflected in the class attribute documentation. The attribute is initialized to None in the constructor and update to a list of all False booleans in the import_alignment function. The generate_sub_alignment method also subsets this attribute so that the state of all sequences will be tracked properly. Finally, import_alignment was updated so that it checks whether the query sequence was found or not, if not an exception is raised. Relaxed the query_id requirements, it now no longer has to be preceded by >query_ to be considered a proper query ID, whatever the users passes in will be used as the sequence id. Methods referencing the query id which expected it to include the leading carrot (>) have been updated so they just compare against the query_id without additional indexing.