Created AlignmentDistanceCalculator class and initial tests.
This class is meant to remove some of the bloat found in the SeqAlignment class and mirror the scoring done in Biopython and ETC better. The first set of tests written for the class ensure that the calculation for identity and one scoring matrix (blosum62) provide the same distance scores as those found by biopython's distance_calculator in the Bio.Phylo.Tree_Construction package. The code is a faster update of what the biopython code already does (I have plans to submit this as an update to the biopython project. There is an additional method to calculate the distance scores produced by ETC during its calculation, but this still needs to be tested.