Continued working on get_et_distance_small test
Corrected the identity computation (removed the 1-id score) such that the score from the Python implementation and the ETC implementation are much closer. They are however still higher than I expected. I would like to check next if the gap distances are the same for the two methods since that is one place I can imagine differences cropping up.