Creating a DataSetGenerator
First draft of all methods going from reading in the list of pdbs, downloading them, extracting their sequences, blasting them, identifying the filtered sequence pileup, and aligning them in both the fa and msf formats. I would like to change the methods so that they are id specific and do not all iterate over the full list of ids.