Refining, documenting, and testing DataSetGenerator
I spent time refining the functions in the data set generator to have meaningful returns and also to check whether a step has already been performed and load the data if possible. I also tried to document the code thoroughly. I have added a function outside the class definition to customize the uniref fasta files by removing low quality proteins and fragments. Finally, I have begun testing on the class but it is not yet complete.