Continued working on DataSetGenerator documentation and testing
Added documentation and testing for build_dataset, as well as ensuring that all of the variables for all of the private methods are used correctly. Separated out a function for identifying which identity bin a sequence belongs in so that this code does not have to be duplicated in the _restrict_sequences function. This function has not yet been tested. Added data loading to the _restrict_sequences method and abstracted out some of the repeated code. Added documentation and testing for the _align_sequences method. In addition I added the option to produce one or the other type of alignment and to check if the alignment files already exist or not. Updated the documentation for batch_iterator and filter_uniref_fasta.