Corrected bin determination for DataSetGenerator
Incorrect handling of the identity calculation was causing all sequences to be filtered out, this has been corrected now. In addition I made sure to create the Alignments directory is created in the _align_sequences function because MUSCLE was throwing errors trying to write to a non-existant directory. All tests pass at the moment, however it is pretty obvious that loading in the _restrict_sequences function is not working correctly (as evidenced by errors when re-running tests without clearing the old data).