Added test for ignore filter size in DataSetGenerator and started testing AlignmentDistanceCalculator again.
New test for DataSetGenerator checks if no pileup was created with given settings and if so uses the ignore filter size option. Also created setUpClass and tearDownClass methods for test_AlignmentDistanceCalculator to make use of the DataSetGenerator. Changed the setUp and tearDown methods to work with the class level ones.