Started writing PhylogeneticTree class
This class is meant to mirror the capabilities of parts of the ETC code base as well as remove bloat from the SeqAlignment class. So far I have implemented methods for calculating distance matrices and building trees. The agglomerative tree building method depends on code scribbed from StackOverflow for converting an agglomerative clustering tree to an nhx tree. I probably need to find a way to convert the nhx trees read in (from ETC) to rooted trees since they are not rooted by default. Next steps are to find a good way to traverse the tree. A top down traversal would allow for easy naming of the nonterminal nodes to match the ETC naming conventions. A bottom up traversal would allow for an intuitive method of determining subalignments, position specific amino/nucleic acid frequencies, and pair specific amino/nucleic acid frequencies.