Updated DataSetGenerator to add Alignemnt Characterization and command line tools.
Added a new method which uses the same methods as _restrict_sequencs to characterize sequences in a fasta formatted alignment. This method now has a command line tool enabled by argparse. I also created command line tools for filtering the Uniref fasta as well as for building a data set.