Update DataSetGenerator _restrict_sequences
This method now returns all sequences falling within the expected identity bins. It still returns a final identity bin as the first return, but this no longer tracks the last bin which contributed, but rather the first bin where 125 sequences (cumulative) were found. Also corrected the call to the method which identifies the similarity bin since I made it a module level method instead of a class level method.