Changed alignment method in DataSetGenerator to clustalw
The changed alignment method still has the ability to produce fasta and/or msf formatted alignment files. If both are selected the alignment is run only once to produce the fasta alignment and then converted into the msf format. Either option can be turned off resulting in one alignment to produce the desired format. Also added a check at the beginning to make sure that the appropriate pileup file had already been generated.