Updating DataSetGenerator to match previous lab methods
Made importing the protein list a module level method instead of a class method. It now takes the full file path to a protein list and imports five letter (not four letter) PDB id codes (where the first four characters are the PDB id and the last character is the chain specifier). The return is a dictionary of dictionaries where the first level key is the PDB id and the second level dictionary contains only the key 'Chain' and the parsed in chain specifier. The expectation is that this will be filled in further as the data set is constructed.