Updating DataSetGenerator to match previous lab methods
Made parsing the query sequence from a pdb file a module level method instead of a class method. It now takes the protein id (PDB id) the chain identifier, path to a directory where sequence data can be stored, and path to a downloaded protein structure. The return is the file path where the structure sequence was stored in fasta format.