Updating DataSetGenerator to match previous lab methods
Made BLASTing the query sequence from a fasta file a module level method instead of a class method. It now takes the protein id (PDB id), path to a directory where BLAST data can be stored, and path to a sequence fasta file, as well as BLAST parameters. The return is the file path where the BLAST data was written in xml format.