Updating DataSetGenerator to match previous lab methods
Made filtering BLAST results a module level method instead of a class method. I renamed the function so it is no longer restrict_sequences, it is now filter_blast_sequences. It now takes the protein id (PDB id), path to a directory where filtered sequence data can be stored, the path to a BLAST xml file, as well as filtering parameters (). The returns are the sequence count after filtering and the file path where the filtered sequences were written in fasta format.