Updating DataSetGenerator to match previous lab methods
Made the align sequences method a module level method instead of a class method. It now takes a protein id, path to a directory where alignments can be written, pileup file path, and some booleans for which file types to return. The return is the filename (or none) for each of the files produced.