Updating DataSetGenerator to match previous lab methods
Made the identity filter method a module level method instead of a class method. It now takes a protein id, path to a directory where filtered alignments can be written, alignment file path, and the max identity value sequences can have. The return is the count of sequences which passed the filter and the filename for the filtered alignment.