Finished first draft of build data set method for new module structure.
This still needs to be tested, but I think it is a good first draft. To speed things up I think the whole thing can benefit from multiprocessing (though there need to be some smart blocking behaviors added so that we don't take advantage of more cores than are available since BLAST may be called with many threads.