Working on PhylogeneticTree traversal/iter algorithm
Cleaned up the method I use to travers trees in SeqAlignment since I think it would be intuitive to turn that into an iter method. Corrected the branch length used to sort nodes, should be total_branch_length not just the branch_length between a parent and child node. Also corrected the spelling of building in PhylogenetiTree.