PhylogeneticTree: Tree Traversals
I think I am very close to having the tree traversal that I want, however at the moment it is not obvious to me whether the UPGMA tree naming agrees with what I expect or not (it appears the branch lengths in the images created by Bio.Phylo.draw are not to scale with what is actually in the tree? or that I am computing cumulative branch length incorrectly).