Troubleshooting PhylogeneticTree traversal
Still do not have quite the behavior I want based on the node labels generated by ETC or by the UPGMA Bio.Phylo methods. I am still not sure if this is because they do not behave the way I think they should or if this is because the naming conventions by those implementations are different than I expected. Did add code to the upgma and custom tree imports to name the tree nodes in ascending order as they reach the bottom as opposed to descending order (upgma) and with strings as opposed to integers (ETC) since that is interpreted as confidence not node name when importing the NHX tree.