Troubleshooting, Documenting, and Testing PhylogeneticTree
Added significantly more documentation to the existing PhylogeneticTree class. Also pared it down so that it does not duplicate the functionality of other classes but takes in their results as input. Corrected the conversion of the agglomerative clustering tree produced by sklearn to the Newick format. It now behaves correctly and also provides stable labels to the internal nodes. Tests have been added for init as well as construct tree with the current three tree building methods. Still need to: 1) Compare the tree constructed by get_et_distance and upgma tree building to the ETC tree, 2) complete and test a traversal algorithm that emits nodes an a way that is intuitive for trace/cluster construction, 3) write a test for tree serialization. Might also be nice to write a method which will homogenize the internal node naming across all of the tree construction methods.