Made corrections to DataSetGenerator to handle larger data set.
The changes include attempting a PDB download a second time as an obsolete file, and returning None if that also does not succeed. This is followed by a check when calling the parse function on the PDB file so that parsing does not get attempted if None was returned. Also added a check so that if the specified chain ('A'  for all the structures in our datasets), is not available, the first (alphabetical) chain is used and the 'Chain' field in the protein_data dict for the protein is updated.