Created new utils file
This file is meant to hold functions which might be needed in a variety of different classes to perform some basic function which is not necessarily better suited to any one class. At the moment I have added the build_mapping and convert_seq_to_numeric functions. These were originally part of the AlignmentDistanceCalculator but are also needed in other classes like the SeqAlignment class.