Created an explicit Alphabet object for the DNA and protein alphabets used in the ET code base.
This is interesting because it actually mimics what is used by the Bio.Phylo.TreeConstruction.DistanceCalculator to a large degree but that class does not use an Alphabet object but rather uses an explicit string of letters. I also found that many of the predefined Alphabet classes do not have size attributes that match their letters attribute and that apparently the Alphabet class and its sub-classes do not perform inheritance in a way that you can check if something is an Alphabet object.