Lots of changes to SeqAlignmnent
Commented out and/or removed a lot of the code which is now being taken care of by other classes. Left a few methods which ar not appropriate for the SeqAlignment class because I do not want to forget to implement them elsewhere. Added a new method to create a consensus sequence, doing my best to implment a majority scheme as opposed to the hard percentile cut off scheme used by Bio.Align.AlignInfo.SummaryInfo in the dumb_consensus and gap_consensus methods. Alo added a method to prune an alignment based on the gap content of the sequences which is part of the ETC software. This required two new methods one to compute z-scores for gaps and one two compute deviation from the consensus sequence (in terms of gaps).