Updating SeqAlignment and tests
Updated several more tests for the class, commenting out and moving methods according to what I expect to keep and cut by the end of this edit. Also updated any method which returns an alignment, such that it returns an alignment where the attribute values are deep copies not references to the attributes of the original.