Performed an extensive overhaul of the SeqAlignment class and its testing.
A number of the tests I had written failed once they were actually run. This commit is a result of troubleshooting those failures. Many of the corrections were a result of typos or careless mestakes having to do with the switch from hard coded static examples (pre-generated alignment files), to the DataSetGenerator which builds the example input. A few errors had to do with consistency, in some methods the query)sequence of the SeqAlignment class was passed as a Seq object and in other cases it was passed as a SeqRecord. Another issue was that I failed to update a lot of the code in the AlignmentDistanceCalculator which depended on functions which are now in the utils file, those have been corrected (some of the fixes rely on calling private attributes like MultipleSequenceAlignment._alphabet, hopefully this will continue to be allowed, since at the moment there are no geetters for this attribute).